requirements.txt
numpy
scipy
matplotlib
pyserial
pandas
scikit-learn
filterpy
-e .
